make TestEnv data generation platform independent (32/64bit)

+++ test/model/testenv.go
@@ -80,7 +80,7 @@ func (env *TestEnv) Populate(count uint) {
 	var entities = make([]*Entity, toInsert)
 	var i = uint(0)
 	for coef := -limit; i < toInsert; coef += step {
-		entities[i] = entity47(int(coef))
+		entities[i] = entity47(int64(coef))
 		i++
 	}
 
@@ -105,7 +105,7 @@ func Entity47() *Entity {
 }
 
 // create a test entity ("47" because int fields are multiples of 47)
-func entity47(coef int) *Entity {
+func entity47(coef int64) *Entity {
 	// NOTE, it doesn't really matter that we overflow the smaller types
 	var Bool = coef%2 == 1
 
@@ -117,12 +117,12 @@ func entity47(coef int) *Entity {
 	}
 
 	return &Entity{
-		Int:          47 * coef,
+		Int:          int(int32(47 * coef)),
 		Int8:         47 * int8(coef),
 		Int16:        47 * int16(coef),
 		Int32:        47 * int32(coef),
 		Int64:        47 * int64(coef),
-		Uint:         47 * uint(coef),
+		Uint:         uint(uint32(47 * coef)),
 		Uint8:        47 * uint8(coef),
 		Uint16:       47 * uint16(coef),
 		Uint32:       47 * uint32(coef),
